add Qt6 support and use in static binaries
- update CMakeLists to work with Qt5.15 / Qt6
- add `-DSME_WITH_QT6=on` bool flag to CMake options
- CI
  - sme_deps -> 2021.05.11 (includes qt6)
  - linux builds: ubuntu 16.04 -> 18.04
  - linux build: add `-DOpenGL_GL_PREFERENCE=LEGACY` to avoid Qt6 picking up unnecessary libOpenGL dependency
  - linux code quality builds: use gcc instead of clang, since they use static libs built with gcc
  - manylinux images -> 2021.05.12
  - python 3.5 is EOL, stop building python 3.5 wheels
- resolves #528

+++ src/gui/widgets/qplaintextmathedit.cpp
@@ -412,7 +412,7 @@ static std::pair<int, bool> getClosingBracket(const QString &expr, int pos,
                                               int sign) {
   int len = 0;
   int count = sign;
-  int iEnd = (sign < 0) ? -1 : expr.size();
+  int iEnd = (sign < 0) ? -1 : static_cast<int>(expr.size());
   for (int i = pos + sign; i != iEnd; i += sign) {
     ++len;
     if (expr[i] == ')') {
